+++ src/index.ts
@@ -31,7 +31,6 @@ import * as CheckinGateFactory from './factory/place/checkinGate';
 import * as MovieTheaterFactory from './factory/place/movieTheater';
 import PriceCurrency from './factory/priceCurrency';
 import * as PropertyValueFactory from './factory/propertyValue';
-import * as ReservationFactory from './factory/reservation';
 import * as EventReservationFactory from './factory/reservation/event';
 import { ReservationStatusType } from './factory/reservationStatusType';
 import SortType from './factory/sortType';
@@ -121,9 +120,9 @@ export import person = PersonFactory;
 export import personType = PersonType;
 export import propertyValue = PropertyValueFactory;
 export namespace reservation {
-    export import reservationType = ReservationFactory.ReservationType;
     export import event = EventReservationFactory;
 }
+export import reservationType = chevre.reservationType;
 export import sortType = SortType;
 export import stock = StockFactory;
 export namespace place {