+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 予約インターフェースをChevreを継承するように拡張
+- install @chevre/factory@1.0.0-alpha.88
+- install @cinerino/factory@1.0.0-alpha.126
 
 ### Deprecated